Header file cleanup, backported from [bug6e8afe516d-87] branch

+++ generic/tkCanvas.c
@@ -1027,7 +1027,7 @@ CanvasWidgetCmd(
 		result = TCL_ERROR;
 		goto done;
 	    }
-	    if (mask & (unsigned long) ~(ButtonMotionMask|Button1MotionMask
+	    if (mask & (unsigned) ~(ButtonMotionMask|Button1MotionMask
 		    |Button2MotionMask|Button3MotionMask|Button4MotionMask
 		    |Button5MotionMask|ButtonPressMask|ButtonReleaseMask
 		    |EnterWindowMask|LeaveWindowMask|KeyPressMask
@@ -4747,7 +4747,7 @@ CanvasBindProc(
     XEvent *eventPtr)		/* Pointer to X event that just happened. */
 {
     TkCanvas *canvasPtr = clientData;
-    unsigned long mask;
+    int mask;
 
     Tcl_Preserve(canvasPtr);
